Fixed property data type

+++ Project-2.Data/Interfaces/IPropertyRepository.cs
@@ -3,6 +3,6 @@ using Project_2.Models;
 namespace Project_2.Data;
 
 public interface IPropertyRepository : IBaseRepository<Property> {
-    public Task<IEnumerable<Property>> GetAllWithFilters(decimal priceMin, decimal priceMax, int numBedroom, double numBathroom);
+    public Task<IEnumerable<Property>> GetAllWithFilters(decimal priceMin, decimal priceMax, int numBedroom, float numBathroom);
     public void Update(Property propertyInfo);
 }
